// src/app/enums/message.ts
/**
 * Enumeration with the messages for errors
 */
export enum Message {
  NOT_FOUND_ASSIGNMENT = 'Se esperaba ~',
  NOT_FOUND_DATA_TYPE = 'No está especificado el tipo de dato',
  NOT_FOUND_EXPRESSION = 'Se esperaba una expresión',
  NOT_FOUND_FUNCTIONS = 'Esta clase no tiene funciones',
  NOT_FOUND_IDENTIFIER = 'Se esperaba un identificador',
  NOT_FOUND_LEFT_BRACKET = 'Se esperaba [',
  NOT_FOUND_LEFT_PARENTHESIS = 'Se esperaba (',
  NOT_FOUND_LOGIC_OPERATOR = 'Se esperaba un operador logico',
  NOT_FOUND_ARITHMETIC_OPERATOR = 'Se esperaba un operador arimetico',
  NOT_FOUND_ARITHMETIC_VALUE = 'Se esperaba un valor arimetico',
  NOT_FOUND_LOGIC_VALUE = 'Se esperaba un valor logico',
  NOT_FOUND_NEXT_PARAM = 'Se esperaba un nuevo parametro separado por § ó )',
  NOT_FOUND_RETURN = 'No está especificado el tipo de retorno',
  NOT_FOUND_RIGHT_BRACKET = 'Se esperaba ]',
  NOT_FOUND_RIGHT_PARENTHESIS = 'Se esperaba )',
  NOT_FOUND_SENTENCES = 'El bloque de setencias está vacío',
  NOT_FOUND_TWO_POINT_OPERATOR = 'Se esperaba æ para especificar el tipo de retorno',
  NOT_FOUND_WHEN = 'Se esperaba `when`',
  NOT_FOUND_END_LINE_OPERATOR = 'Se esperaba ■',
}