// src/app/enums/reserved-word.ts
/**
 * Enumeration with the reserved words
 */
export enum ReservedWord {
  ARIT = 'arit',
  ASSIGNMENT = 'assignment',
  ADJUST = 'adjust',
  ADJUST_INIT = 'adjustInit',
  BOOLEAN = 'boolean',
  BREAK = 'break',
  CHAR = 'char',
  CONTINUE = 'continue',
  DECIMAL = 'decimal',
  END_PROJECT = 'EndProject',
  FALSE = 'false',
  FUNCTION = 'function',
  INT = 'int',
  LOG = 'log',
  IMMUTABLE = 'InMut',
  MUTABLE = 'Mut',
  OTHER = 'other',
  REL = 'rel',
  PRIVATE = 'private',
  PUBLIC = 'public',
  RETURN = 'return',
  START_PROJECT = 'StrProject',
  STRING = 'string',
  TRUE = 'true',
  VOID = 'void',
  WHEN = 'when',
}